+++ dev/test/test_panda/fake_real_panda.py
@@ -3,7 +3,7 @@
 
 import pytest
 from perls2.utils.yaml_config import YamlConfig
-from perls2.ros_interfaces.redis_keys import *
+from perls2.redis_interfaces.redis_keys import *
 from perls2.robots.real_panda_interface import RealPandaInterface
 
 class FakePandaInterface(RealPandaInterface):
