+++ dev/test/test_panda/test_panda_nuc.py
@@ -10,14 +10,14 @@ This script is meant to be run on the nuc, may be extended to ws.
 """
 import redis
 import pytest
-import numpy as np 
+import numpy as np
 import time
-from perls2.ros_interfaces.redis_keys import *
-from perls2.ros_interfaces.redis_values import * 
+from perls2.redis_interfaces.redis_keys import *
+from perls2.redis_interfaces.redis_values import *
 from perls2.ros_interfaces.panda_ctrl_interface import PandaCtrlInterface
 from perls2.robots.real_panda_interface import RealPandaInterface
 
-from perls2.ros_interfaces.redis_interface import PandaRedisInterface
+from perls2.redis_interfaces.redis_interface import PandaRedisInterface
 from perls2.utils.yaml_config import YamlConfig
 
 from dev.test.test_panda.fake_franka_panda import FakeFrankaPanda
@@ -58,7 +58,7 @@ def test_get_driver_state_model(fake_driver, panda_ctrl):
 def test_update_model(fake_driver, panda_ctrl):
 	fake_driver.set_fake_state()
 	panda_ctrl.update_model()
-	assert len(panda_ctrl.model.ee_pos) == 3 	
+	assert len(panda_ctrl.model.ee_pos) == 3
 	assert panda_ctrl.model.ee_ori_mat.shape == (3,3)
 	assert panda_ctrl.model.ee_ori_quat.shape == (4,)
 	assert panda_ctrl.model.joint_pos.shape == (7,)
@@ -75,7 +75,7 @@ def test_step(fake_driver, panda_ctrl, real_panda):
 	panda_ctrl.update_model()
 	panda_ctrl.controller = panda_ctrl.make_controller_from_redis(panda_ctrl.get_control_type(), panda_ctrl.get_controller_params())
 
-	curr_ee_pos = panda_ctrl.model.ee_pos 
+	curr_ee_pos = panda_ctrl.model.ee_pos
 	curr_ee_ori = panda_ctrl.model.ee_ori_quat
 
 	real_panda.set_ee_pose(set_pos=curr_ee_pos, set_ori=curr_ee_ori)
