+++ dev/test/test_panda/test_panda_ws.py
@@ -15,7 +15,7 @@ import numpy as np
 from perls2.utils.yaml_config import YamlConfig
 from perls2.robots.real_panda_interface import RealPandaInterface
 from perls2.ros_interfaces.panda_ctrl_interface import PandaCtrlInterface
-from perls2.ros_interfaces.redis_keys import *
+from perls2.redis_interfaces.redis_keys import *
 
 from dev.test.test_panda.fake_franka_panda import FakeFrankaPanda
